The previous version of the StatefulTokenizer was not thread-safe, it is now.
